respond to Jeremy's feedback
