+++ packages/math-input/src/components/input/mathquill-instance.ts
@@ -1,5 +1,5 @@
 import MathQuill from "mathquill";
 
-import {MathQuillInterface} from "../../types";
+import {MathQuillInterface} from "./mathquill-types";
 
 export default MathQuill.getInterface(2) as MathQuillInterface;
